Revert "Update base call filter"
This reverts commit 138e8924195640f4916d83534b79313cbc1a3be6.
